use 1 file for duration_in_seconds
