+++ litellm/proxy/management_endpoints/team_callback_endpoints.py
@@ -90,8 +90,8 @@ async def add_team_callbacks(
     """
     try:
         from litellm.proxy.proxy_server import (
-            _duration_in_seconds,
             create_audit_log_for_update,
+            duration_in_seconds,
             litellm_proxy_admin_name,
             prisma_client,
         )
